requirements.txt
streamlit==1.35.0
openai==1.30.1
easyocr==1.7.1
Pillow==10.3.0
numpy==1.26.4
googletrans==4.0.0-rc1
python-dotenv==1.0.1
